Fix assembler "unterminated string" error when a string literal ended with "\\"
Since escape sequences are handled *after* the full lexeme is scanned, the initial scan didn't understand that the final backslash wasn't for the end quote and was instead escaped by the previous backslash. A better fix in the future may be to scan the lexeme and the literal at the same time.